+++ Runtime/Core/Fields/AbstractField.cs
@@ -17,10 +17,25 @@ namespace ActionCode.UISystem
             set
             {
                 if (value.Equals(this.value)) return;
+
+                if (IsReadOnly)
+                {
+                    DenyReadOnlyValueChange();
+                    return;
+                }
+
                 this.value = value;
                 ChangeValue(this.value);
             }
         }
+
+        /// <summary>
+        /// Whether the value is read only and cannot be modified.
+        /// <para>
+        /// <see cref="OnReadOnlyValueChangeDenied"/> event will be fired when trying to change read only values.
+        /// </para>
+        /// </summary>
+        public bool IsReadOnly { get; set; }
 
         /// <summary>
         /// Event fired when the Value is changed.
@@ -29,6 +44,11 @@ namespace ActionCode.UISystem
         /// It will not be fired if the value is set using <see cref="SetValueWithoutNotify(T)"/>.
         /// </remarks>
         public event Action<T> OnValueChanged;
+
+        /// <summary>
+        /// Event fired when trying to change a read only Value, when <see cref="IsReadOnly"/> is true.
+        /// </summary>
+        public event Action OnReadOnlyValueChangeDenied;
 
         private T value;
 
@@ -39,5 +59,6 @@ namespace ActionCode.UISystem
         public virtual void SetValueWithoutNotify(T value) => this.value = value;
 
         protected virtual void ChangeValue(T value) => OnValueChanged?.Invoke(value);
+        protected virtual void DenyReadOnlyValueChange() => OnReadOnlyValueChangeDenied?.Invoke();
     }
 }